+++ packages/drawer/src/Component.screenshots.test.tsx
@@ -1,0 +1,30 @@
+import { setupScreenshotTesting, createSpriteStorybookUrl } from '../../screenshot-utils';
+
+const screenshotTesting = setupScreenshotTesting({
+    it,
+    beforeAll,
+    afterAll,
+    expect,
+});
+
+describe(
+    'Drawer',
+    screenshotTesting({
+        cases: [
+            [
+                'sprite',
+                createSpriteStorybookUrl({
+                    componentName: 'Drawer',
+                    size: { width: 350, height: 400 },
+                    knobs: {
+                        open: [false, true],
+                        children: 'Content',
+                    },
+                }),
+            ],
+        ],
+        screenshotOpts: {
+            fullPage: true,
+        },
+    }),
+);