+++ packages/icon-button/src/Component.tsx
@@ -63,12 +63,11 @@ export const IconButton = forwardRef<HTMLButtonElement, IconButtonProps>(
                     colorStyles[colors].loader,
                 )}
                 size='s'
-                leftAddons={
-                    <span className={cn(styles.iconWrapper, styles[size])}>
-                        <Icon className={styles.icon} />
-                    </span>
-                }
-            />
+            >
+                <span className={cn(styles.iconWrapper, styles[size])}>
+                    <Icon className={styles.icon} />
+                </span>
+            </Button>
         );
     },
 );