+++ packages/select/src/components/select-mobile/Component.tsx
@@ -1,5 +1,6 @@
 import React, { ReactNode } from 'react';
 
+import { BottomSheetProps } from '@alfalab/core-components-bottom-sheet';
 import { BaseSelectMobile } from '../base-select-mobile';
 
 import { Field as DefaultField } from '../field';
@@ -12,14 +13,20 @@ import { BaseSelectProps } from '../../typings';
 export type SelectMobileProps = Omit<BaseSelectProps, 'OptionsList' | 'Checkmark' | 'onScroll'> & {
     /**
      * Футер
+     * @deprecated Используйте bottomSheetProps.actionButton
      */
     footer?: ReactNode;
 
     /**
      * Будет ли свайпаться шторка
-     *
+     * @deprecated Используйте bottomSheetProps.swipeable
      */
     swipeable?: boolean;
+
+    /**
+     * Дополнительные пропсы шторки
+     */
+    bottomSheetProps?: Partial<BottomSheetProps>;
 };
 
 export const SelectMobile = ({