+++ packages/calendar-with-skeleton/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.3](https://github.com/core-ds/core-components/compare/@alfalab/core-components-calendar-with-skeleton@4.2.2...@alfalab/core-components-calendar-with-skeleton@4.2.3) (2022-06-20)
+
+
+### Bug Fixes
+
+* **calendar:** increased specificity of the button css class ([#103](https://github.com/core-ds/core-components/issues/103)) ([3d20c11](https://github.com/core-ds/core-components/commit/3d20c112a1c161951afa2e1c6fca8d3c048501e7))
+
+
+
+
+
 ## [4.2.2](https://github.com/core-ds/core-components/compare/@alfalab/core-components-calendar-with-skeleton@4.2.1...@alfalab/core-components-calendar-with-skeleton@4.2.2) (2022-06-03)
 
 **Note:** Version bump only for package @alfalab/core-components-calendar-with-skeleton